@@ -3,7 +3,7 @@ cmake_minimum_required(VERSION 3.14)
 project(snaildb CXX)
 
 set(CMAKE_EXPORT_COMPILE_COMMANDS True)
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED True)
 
 include(FetchContent)
@@ -26,7 +26,7 @@ add_library(core "libs/core/db.cc")
 target_include_directories(core PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/libs/lsm_tree/")
 target_compile_options(core PRIVATE -g)
 
-add_library(lsm_tree "libs/lsm_tree/cbf.h" "libs/lsm_tree/lsm.cc" "libs/lsm_tree/sst.cc" "libs/lsm_tree/segment.cc")
+add_library(lsm_tree "libs/lsm_tree/cbf.h" "libs/lsm_tree/lsm.cc" "libs/lsm_tree/sst.cc" "libs/lsm_tree/segment.cc" "libs/lsm_tree/record.h")
 target_include_directories(lsm_tree PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/libs/lsm_tree/" "${CMAKE_CURRENT_SOURCE_DIR}/libs/util/")
 target_compile_options(lsm_tree PRIVATE -g)
 
@@ -35,15 +35,9 @@ target_include_directories(snaildb PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/libs/lsm
 target_link_libraries(snaildb PRIVATE core lsm_tree spdlog::spdlog)
 target_compile_options(snaildb PRIVATE -g)
 
-add_executable(snail-test "test/test.cc")
+add_executable(snail-test "test/behaviour_tests.cc")
 target_include_directories(snail-test PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/libs/lsm_tree/" "${CMAKE_CURRENT_SOURCE_DIR}/libs/util")
 target_link_libraries(snail-test PRIVATE core lsm_tree Catch2::Catch2WithMain)
 target_compile_options(snail-test PRIVATE -g)
 
-#Auto run tests. Taken from: https://stackoverflow.com/a/32902353
-add_custom_command(
-  TARGET snail-test
-  COMMENT "Run tests"
-  POST_BUILD
-  COMMAND snail-test
-  )
+add_test(NAME snail-test COMMAND snail-test)
